Added support for generating processor cores rather than full system models. (#316)
Move toKamiRes to TlUh.v and change response type to Device.Res.
Create an equivalent generator for just processorPipeline. And create an instance of the pipeline in Instance.v
Create a command line option in the Verilog Target.hs file.an option that pretty prints verilog just for the Pipeline (processor core without device models).
Add a command line parameter to doGenerate so that it fills in Target.hs with the Pipeline verilog and the doGenerate stops at verilog generation - i.e. does not do verilator or make.
The goal is to have a command line script that generates the verilog model of the processor core without devices, etc.